rename data migrations for clarity
