+++ packages/db/prisma/data-migrations/index.ts
@@ -1,16 +1,16 @@
 // codegen:start {preset: barrel, include: ./**/*.ts, exclude: ./**/!*.ts}
-export * from './2023-03-29'
-export * from './2023-03-31'
-export * from './2023-04-04/index'
-export * from './2023-04-04b'
-export * from './2023-04-05'
-export * from './2023-04-05b/index'
-export * from './2023-04-06/index'
-export * from './2023-04-10/index'
-export * from './2023-04-10b'
-export * from './2023-04-11'
-export * from './2023-04-11b'
-export * from './2023-04-21'
-export * from './2023-04-25/index'
-export * from './2023-04-26/index'
+export * from './2023-03-29_nest-attribs-primary-cntry'
+export * from './2023-03-31_suggestion-source'
+export * from './2023-04-04_badges/index'
+export * from './2023-04-04_single-loc-org-link'
+export * from './2023-04-05_user-title-namespace'
+export * from './2023-04-05_website-lang-codes/index'
+export * from './2023-04-06_attrib-updates/index'
+export * from './2023-04-10_phone-descriptions/index'
+export * from './2023-04-10_social-media-services'
+export * from './2023-04-11_basic-phone-types'
+export * from './2023-04-11_permissions-attributes'
+export * from './2023-04-21_badge-render-field'
+export * from './2023-04-25_attr-supp-schemas/index'
+export * from './2023-04-26_attr-supp-interpolation/index'
 // codegen:end
